Extract useDeployMutation hook from DeployTab
Moves isRedeploy detection, the triggerDeployment mutation, and the
onDeploy handler into a dedicated hook, removing the largest single
logic block from DeployTab.
